barbuzz/backend/requirements.txt
# Core packages
django>=5.0.0
djangorestframework>=3.14.0
django-allauth>=0.60.0
django-cors-headers>=4.3.1
django-ninja >= 1.3.0,<2
dj-rest-auth >= 7.0.1

# API documentation
drf-spectacular>=0.27.0

# Database
psycopg2-binary>=2.9.9  # For PostgreSQL

# google api
google-api-python-client>=2.0.0
googlemaps >=4.10.0
